backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
amqp==5.3.1
    # via kombu
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
asttokens==3.0.0
    # via stack-data
billiard==4.2.1
    # via celery
boto3==1.35.74
    # via -r requirements.in
botocore==1.35.74
    # via
    #   boto3
    #   s3transfer
celery==5.4.0
    # via
    #   -r requirements.in
    #   django-celery-beat
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.5
    # via django-celery-beat
decorator==5.1.1
    # via ipython
dj-database-url==2.3.0
    # via -r requirements.in
dj-rest-auth==7.0.0
    # via -r requirements.in
django==4.2.16
    # via
    #   -r requirements.in
    #   dj-database-url
    #   dj-rest-auth
    #   django-celery-beat
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-reversion
    #   django-role-permissions
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   graphene-django
django-celery-beat==2.7.0
    # via -r requirements.in
django-cors-headers==4.6.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==24.3
    # via -r requirements.in
django-reversion==5.1.0
    # via -r requirements.in
django-role-permissions==3.2.0
    # via -r requirements.in
django-storages==1.14.4
    # via -r requirements.in
django-timezone-field==7.0
    # via django-celery-beat
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   dj-rest-auth
exceptiongroup==1.2.2
    # via ipython
executing==2.1.0
    # via stack-data
graphene==3.4.3
    # via graphene-django
graphene-django==3.2.2
    # via -r requirements.in
graphql-core==3.2.5
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==3.2.0
    # via
    #   graphene
    #   graphene-django
ipython==8.30.0
    # via -r requirements.in
jedi==0.19.2
    # via ipython
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.4.2
    # via celery
matplotlib-inline==0.1.7
    # via ipython
parso==0.8.4
    # via jedi
pendulum==3.0.0
    # via -r requirements.in
pexpect==4.9.0
    # via ipython
pprintpp==0.4.0
    # via -r requirements.in
promise==2.3
    # via graphene-django
prompt-toolkit==3.0.48
    # via
    #   click-repl
    #   ipython
psycopg==3.2.3
    # via -r requirements.in
psycopg-binary==3.2.3
    # via psycopg
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.18.0
    # via ipython
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   botocore
    #   celery
    #   graphene
    #   pendulum
    #   python-crontab
    #   time-machine
raven==6.10.0
    # via -r requirements.in
s3transfer==0.10.4
    # via boto3
six==1.16.0
    # via
    #   promise
    #   python-dateutil
sqlparse==0.5.2
    # via django
stack-data==0.6.3
    # via ipython
text-unidecode==1.3
    # via graphene-django
time-machine==2.16.0
    # via pendulum
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via
    #   asgiref
    #   dj-database-url
    #   graphene
    #   ipython
    #   psycopg
tzdata==2024.2
    # via
    #   celery
    #   django-celery-beat
    #   kombu
    #   pendulum
urllib3==2.2.3
    # via botocore
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
